Refactor code that should not be executed upon source
Some files contain code that is not intended to be executed when the
files are sourced, like the code that generates the word CSV file (which
takes minutes to run), and the code that creates a table output.

Because we will source some of these files in the final Shiny app, and
we don't want those pieces of code mentioned above to execute, we need to
modify them by either moving them into a function, so they will only be
executed when we call the function, or simply removing them.